+++ app/pb_kits/playbook/pb_popover/popover.rb
@@ -11,10 +11,6 @@ module Playbook
       prop :tooltip_id
       prop :offset, type: Playbook::Props::Boolean, default: false
 
-      def popover_offset
-        offset ? "0, 8" : "0, 0"
-      end
-
       def classname
         generate_classname("pb_popover_kit")
       end