+++ CHANGELOG.md
@@ -13,6 +13,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 - Adds ability to give context to the Typeahead kit ([#621](https://github.com/powerhome/playbook/pull/621) @terryfinn @web-kat)
 
+### Fixed
+- Correct Popover offset docs [#611](https://github.com/powerhome/playbook/pull/611)
+
 
 ## [4.3.0] 2020-2-14